Extracted passwordEntered logic to new file
