+++ ui/src/logic/ImportScreen/PasswordManager.ts
@@ -1,0 +1,22 @@
+export const passwordEntered = (
+  event: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>,
+  index: number,
+  passwords: string[],
+  setPasswords: (passwords: string[]) => void
+) => {
+  const pass = event.target.value;
+  const newList = Array.from(passwords);
+  newList[index] = pass;
+  setPasswords(newList);
+};
+
+export const uniquePasswordEntered = (
+  event: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>,
+  passwords: string[],
+  setPasswords: (passwords: string[]) => void
+) => {
+  const pass = event.target.value;
+  const newList = Array.from(passwords);
+  newList.fill(pass);
+  setPasswords(newList);
+};
